Moved Like feature to Like.js to split up Post.js
